Store map as backbone view
This makes the map an object-oriented component.
It will be created inside inside init function that is called when jquery document.ready fires. Once it is initialized it will render itself. It will re-render once the window has been resized, but not before the user finished doing so.
This prevents flickering as some browsers fire the resize event multiple times per second.
@see http://drupalmotion.com/article/debounce-and-throttle-visual-explanation